email_monitor_api/BL/app/requirements.txt
colorama
Flask
Flask-Cors
mysqlclient
pandas
pillow
requests
sqlalchemy
